+++ hosts/hermes/darwin-configuration.nix
@@ -20,8 +20,7 @@
 
   system.defaults.dock.persistent-apps = [
     # Update this when firefox-bin-unwrapped is merged
-    (assert pkgs.firefox.meta.unsupported && pkgs.firefox-bin.meta.unsupported;
-      "/Applications/Firefox.app")
+    (assert pkgs.firefox-bin.meta.unsupported; "/Applications/Firefox.app")
     "/Applications/Ghostty.app"
     "/Applications/1Password.app"
     "${pkgs.vscode}/Applications/Visual Studio Code.app"