+++ modules/cache.nix
@@ -1,10 +1,12 @@
 let
   shared = { inputs, keys, pkgs, lib, ... }: {
-    nix.settings.substituters = [ "https://enzime.cachix.org" ];
+    nix.settings.substituters =
+      [ "https://enzime.cachix.org" "https://cache.clan.lol" ];
     nix.settings.trusted-public-keys = builtins.attrValues {
       inherit (keys.signing) aether chi-linux-builder echo;
 
       "enzime.cachix.org" = keys.signing."enzime.cachix.org";
+      "cache.clan.lol" = keys.signing."cache.clan.lol";
     };
   };
 in {