+++ modules/graphical-minimal.nix
@@ -29,7 +29,7 @@
 
     services.xserver.enable = true;
 
-    hardware.pulseaudio.enable = false;
+    services.pulseaudio.enable = false;
     services.pipewire.enable = true;
     services.pipewire.alsa.enable = true;
     services.pipewire.alsa.support32Bit = true;